PIVars::/Ctx:: drop $PIVars::current, use per-Ctx environment..
For <?include?> support, and also for the possibly upcoming SMP
support, and also because it is a cleaner approach, do not use
a PIVars:: internal stack of PI variable environments (hashes),
but instead turn this environment into a member of Ctx::, where
it belongs.

And let the Ctx:: ctor handle the necessary initialization
according to its needs, instead of having PIVars:: worry about
these things that it does not know about.