Drop $Ctx::curr..
In the ongoing effort to implement <?include?> and be able to
implement true SMP for the convertion step this changeset drops
the global $Ctx::curr field.

The main problem to be able to do so were the support log subs
vv/v/e/panic, which used that field to print file/line
information, and so all uses had to be adjusted accordingly.