Expander:: extended array handling..
This changeset also fixes faulty accesses of PIVars::builtin,
which is and ever was a hash-reference and not a hash.

It also silences some rather testing-only verbose log messages.

And, finally, adds some more usage tests for PI variables.
Before it would have been possible to pass arguments to plain
variables.