Convert "multiline" statements to "paired" ones..
It is in fact possible to <?pre?>xxx<?pre end?>, so that the
"multiline" characterization is not true.

To support this for real, handle <?end?> as a normal PI that need
not occur on a line by itself, and even <?begin?> may now be
followed by actual content, so that the following line is
a complete and completely valid S-Web42 file!

   <?begin?>*Hello*, S-Web42.<?end?>