Expander:: tweak Filter::WIPEEL (w) mode for recursion..
When performing recursion that involves a change in the used
Filter::WIPEEL mode, then that would produce unwanted empty
lines when opening the mode with a PI on a line by itself, and
fail to emit a separating newline when data preceedes the
closing PI, speaking about WIPEEL->no WIPEEL and way back.

So extend the state machine to also keep track of the already
emitted output for the current line; sofar the line output
buffer was checked for content to achieve that but of course
that is flushed by _emit(), which in turn is called when
crossing recursion boundaries.