Expander:: tweak _endl/NL_* handling..
Handling of recursion and end-of-line conditions is a bit hairy,
given that it should be possible to seamlessly include entire
files or do whatever else on a single line on the one hand, but
also add proper separating/final newlines on the other hand,
i.e., do *exactly* what the user writes down.
So this changeset is about newline handling across recursion and
output-emit boundaries.

It also drops the *force* argument to _endl(), and encapsulates
the entire logic in _endl() itself.

And finally it converts ?raw_include? logic to use the normal
_prepare_recursion() / _end_recursion() cycle, which has become
possible with this changeset.