Deep clone Ctx->pivars unless MODE_CONFIGRC..
One remaining bug of S-Web42 was that Ctx->pivars was not deep
cloned, so that, e.g., PIVars::outermost->{TOPMENU}, an array
that is part of the example today, was not stable across file
expansions.

I.e., if file1 modified TOPMENU by adding an entry to this
array, then file2 would see the updated array including
modifications from file1, because even though the entire
PIVars::outermost hash was cloned, the values of its keys were
only reference copied.

Now that S-SymObj offers the new SymObj::clone_ref() support sub
use that on Ctx construction time to actually perform a real
deep clone of Ctx->pivars, unless Ctx::MODE_CONFIGRC is set, in
which case we of course want to even reference copy the hash as
such.