If IGNORE_MODTIME, set modtimes to 0..
Since the user does not trust the modtimes it does not make just
any sense to store them in the cache database, since they are
most likely useless.

But if this is true, then we should not store them at all,
because it can be assumed that they produce larger diffs than
necessary -- set them to 0 to avoid that.