Expander:: tweak ?ifdef?/discard handling..
?ifdef?/?else? uses the discard() mechanism to toggle
"outputability" of data.
That did not work yet if that was turned "on" (i.e. output
"off") in the middle of a line that yet had some valid data on
it.

Add an optional $endl argument to _emit() so that this one can
actually outvote the current discard state, and drive this
invocation properly from within _endl().
This change offers the possibility to not push an "\n" string
onto the data array but leave that task entirely up to _emit()
as necessary.
(And, for the data_condome() cases used by ?perl?/?sh? --
that'll join() the data_condome() in _create_subproc() linewise
anyway, so let that use a "\n" not a '' joiner.)

Also, let ?ifdef?/?else?/?fi? perform an _emit() before the
discard() state is toggled.

Add one more simple minded test case.

While there do some style cleanup.