test.sh: fix for Heirloom sh(1) plus..
Even though ":>file" works for Heirloom sh(1), even in a function
etc., the S-Web42 test.sh fails to run because ":>file" does not
work reliable (rejects to overwrite files).
I got sick after an hour and simply replaced the ":>" with
printf(1), which always works.

While there classify the filenames as being located in the
current directory, just to avoid possible problems with
shells that do not like unqualified paths.